pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zn2-adsorption"
version = "0.1.0"
authors = [
  { name="Scientific Software Engineer" },
]
description = "Zn2+ Adsorption Energy Calculator CLI"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Chemistry",
]
dependencies = [
    "pymatgen>=2024.1.20",
    "ase>=3.22.1",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "ccinput>=1.0.0",
    "rdkit>=2024.9.1",
]

[project.optional-dependencies]
uff = ["openbabel-wheel>=3.1.1"]
rowan = ["rowan-python>=2.1.0", "stjames>=0.0.155"]

[project.scripts]
run-adsorption = "zn2_adsorption.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
